refactor: create new reusuable components for result and resultDetail
